phonebook complete

+++ cpp00/ex01/utils.cpp
@@ -6,7 +6,7 @@
 /*   By: ubuntu <ubuntu@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/06 02:04:07 by ubuntu            #+#    #+#             */
-/*   Updated: 2024/01/14 07:00:31 by ubuntu           ###   ########.fr       */
+/*   Updated: 2024/01/18 03:29:23 by ubuntu           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 std::string	ft_toupper(std::string str)
 {
-	for (int i = 0; i < str.length(); i++)
+	for (long unsigned int i = 0; i < str.length(); i++)
 		str[i] = toupper(str[i]);
 	return str;
 }
